requirements.txt

Django==4.1.7
django-extensions==3.2.1
django-filter==23.1
djangorestframework==3.14.0
requests==2.28.2
shortuuid==1.0.11